requirements.txt
flask
flask-sqlalchemy
werkzeug
python-dotenv
pymysql
cryptography